RAG_Multi_Agent/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
python-multipart
python-dotenv
langchain
langchain-classic
langchain-openai
langchain-community
langchain-text-splitters
pymupdf
python-docx
python-pptx
docx2txt
unstructured
azure-search-documents
azure-identity
